backend/requirements.txt
# ─── Web Framework ───────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ─── Database ────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# ─── Validation ──────────────────────────────────
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# ─── Auth & Security ─────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ─── Redis ───────────────────────────────────────
redis[hiredis]==5.1.0
celery[redis]==5.4.0

# ─── HTTP Client (for ЕСОЗ connector) ────────────
httpx==0.27.2

# ─── File Storage ────────────────────────────────
minio==7.2.9

# ─── PDF Generation ──────────────────────────────
weasyprint==62.3
jinja2==3.1.4

# ─── Email ───────────────────────────────────────
fastapi-mail==1.4.1

# ─── Utils ───────────────────────────────────────
python-dotenv==1.0.1
pytz==2024.2

# ─── Testing ─────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
factory-boy==3.3.1
